Add config file check condition for UNIX setup script

+++ install.sh
@@ -1,6 +1,9 @@
 go install 
 
+if [ ! -e ~/.shavis-go.yaml ]; then
+echo "Copying config file..."
 cp .shavis-go.yaml ~/
+fi
 
 echo "Installation complete!"
 echo "shavis-go config are stored in $HOME/.shavis-go.yaml"
